refactor: move chat route classifier to astra-server-types
