+++ rust/crates/runtime/src/lib.rs
@@ -33,7 +33,6 @@ use axum::{
 };
 use base64::{Engine as _, engine::general_purpose::URL_SAFE};
 use futures_util::StreamExt;
-use regex::RegexBuilder;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use sqlx::{mysql::MySqlPoolOptions, query};
